build time check ohos api version, select dummy source if api version below 21.
